backend/requirements.txt
asyncpg==0.29.0
azure-ai-formrecognizer==3.2.0
azure-identity==1.17.1
azure-storage-blob==12.20.0
fastapi==0.111.1
fuzzywuzzy==0.18.0
openpyxl==3.1.5
pandas==2.2.2
psycopg2-binary==2.9.9
python-Levenshtein==0.26.1
python-dotenv==1.0.1
splink==4.0.2
sqlalchemy==2.0.34
tenacity==9.0.0
uvicorn==0.30.1
requests==2.32.2